@@ -18,8 +18,8 @@
     "prop-types": "^15.5.10"
   },
   "peerDependencies": {
-    "react": ">=16.4.0",
-    "react-dom": ">=16.4.0"
+    "react": ">=16.9.0",
+    "react-dom": ">=16.9.0"
   },
   "devDependencies": {
     "@babel/cli": "^7.1.5",
@@ -41,8 +41,8 @@
     "eslint-plugin-react": "^7.3.0",
     "jest": "^23.1.0",
     "jest-cli": "^23.6.0",
-    "react": "^16.4.0",
-    "react-dom": "^16.4.0",
+    "react": "^16.9.0",
+    "react-dom": "^16.9.0",
     "react-test-renderer": "^15.6.2",
     "rimraf": "^2.6.2",
     "webpack": "^4.26.1",
